feat(toolkit): add `deepCopyArray` function
